Rearrange commit bc4b829 work for our current use
