--- app/helpers/products_helper.rb
@@ -1,2 +1,0 @@
-module ProductsHelper
-end
