+++ app/views/admin/products/edit.html.erb
@@ -1,0 +1,3 @@
+<h1>Editing Product</h1>
+
+<%= render "form", product: @product %>
