--- app/views/products/_product.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! product, :id, :created_at, :updated_at
-json.url product_url(product, format: :json)
