--- app/views/products/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @products, partial: 'products/product', as: :product
