--- app/views/products/new.html.erb
@@ -1,5 +1,0 @@
-<h1>New Product</h1>
-
-<%= render 'form', product: @product %>
-
-<%= link_to 'Back', products_path %>
