--- app/views/products/show.html.erb
@@ -1,4 +1,0 @@
-<p id="notice"><%= notice %></p>
-
-<%= link_to 'Edit', edit_product_path(@product) %> |
-<%= link_to 'Back', products_path %>
