--- app/views/products/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "products/product", product: @product
