+++ config/routes.rb
@@ -4,7 +4,9 @@ Rails.application.routes.draw do
   get 'pages/legal' => "pages#legal"
   get 'pages/shop' => "pages#shop"
 
-  resources :products, except: [:new, :destroy]
+  namespace :admin do
+    resources :products, only: [:index, :edit, :update]
+  end
   resources :companies, only: [:new, :create]
   resources :orders, only: [:new, :create]
   resources :shipping_addresses, only: [:create]
